+++ tkintertools/mpl/__init__.py
@@ -32,5 +32,5 @@ Provides:
 
 from .main import *
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"
 __author__ = "Xiaokang2022 <2951256653@qq.com>"